[v0.87.1][maintainability] Split remote_exec.rs by responsibility
