+++ adl/src/remote_exec.rs
@@ -3,770 +3,44 @@ use std::io::Read;
 use std::time::Duration;
 
 use anyhow::{anyhow, Context, Result};
+#[cfg(test)]
 use base64::Engine;
-use ed25519_dalek::{Signature, Signer, SigningKey, Verifier, VerifyingKey};
 use reqwest::blocking::Client;
 use reqwest::StatusCode;
-use serde::{Deserialize, Serialize};
-use serde_json::{Map, Value};
 use tiny_http::{Header, Method, Response, Server};
 
+#[cfg(test)]
 use crate::adl;
 use crate::provider;
+#[cfg(test)]
 use crate::sandbox;
-use crate::signing;
+
+mod errors;
+mod security;
+mod signing_support;
+mod types;
 
 pub const PROTOCOL_VERSION: &str = "0.1";
 const MAX_REQUEST_BYTES: usize = 5 * 1024 * 1024;
 const REMOTE_REQUEST_SIGNATURE_SCHEMA_V1: &str = "remote_request_signature.v1";
 const REMOTE_REQUEST_SIGNATURE_ALG_ED25519: &str = "ed25519";
 const B64: base64::engine::GeneralPurpose = base64::engine::general_purpose::STANDARD;
-
-/// Client-to-server request for one remotely executed resolved step.
-///
-/// In v0.5 MVP, the local scheduler sends a single step payload to `/v1/execute`.
-/// The remote endpoint does not own scheduling or DAG orchestration.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExecuteRequest {
-    pub protocol_version: String,
-    pub run_id: String,
-    pub workflow_id: String,
-    pub step_id: String,
-    pub step: ExecuteStepPayload,
-    pub inputs: ExecuteInputsPayload,
-    pub timeout_ms: u64,
-    #[serde(default)]
-    pub security: Option<ExecuteSecurityEnvelope>,
-}
-
-/// Resolved step payload executed by the remote endpoint.
-///
-/// The caller provides a fully resolved prompt/provider payload so the server
-/// can execute a single deterministic unit of work.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExecuteStepPayload {
-    pub kind: String,
-    pub provider: String,
-    pub prompt: String,
-    #[serde(default)]
-    pub tools: Vec<String>,
-    pub provider_spec: adl::ProviderSpec,
-    #[serde(default)]
-    pub model_override: Option<String>,
-}
-
-/// Input/state payload snapshot provided with a remote step request.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct ExecuteInputsPayload {
-    #[serde(default)]
-    pub inputs: HashMap<String, String>,
-    #[serde(default)]
-    pub state: HashMap<String, String>,
-}
-
-/// Security envelope attached to remote execution requests.
-///
-/// This metadata is validated centrally before remote execution and provides
-/// deterministic policy gating for trust and sandbox checks.
-#[derive(Debug, Clone, Serialize, Deserialize, Default)]
-pub struct ExecuteSecurityEnvelope {
-    #[serde(default)]
-    pub require_signature: bool,
-    #[serde(default)]
-    pub require_key_id: bool,
-    #[serde(default)]
-    pub signed: bool,
-    #[serde(default)]
-    pub key_id: Option<String>,
-    #[serde(default)]
-    pub signature_alg: Option<String>,
-    #[serde(default)]
-    pub key_source: Option<String>,
-    #[serde(default)]
-    pub request_signature: Option<RemoteRequestSignatureV1>,
-    #[serde(default)]
-    pub allowed_algs: Vec<String>,
-    #[serde(default)]
-    pub allowed_key_sources: Vec<String>,
-    #[serde(default)]
-    pub sandbox_root: Option<String>,
-    #[serde(default)]
-    pub requested_paths: Vec<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-pub struct RemoteRequestSignatureV1 {
-    pub schema_version: String,
-    pub alg: String,
-    #[serde(default)]
-    pub key_id: Option<String>,
-    #[serde(default)]
-    pub public_key_b64: Option<String>,
-    pub sig_b64: String,
-}
-
-/// Remote execution response contract for `/v1/execute`.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ExecuteResponse {
-    pub ok: bool,
-    pub run_id: String,
-    pub workflow_id: String,
-    pub step_id: String,
-    pub result: Option<String>,
-    pub artifacts: Vec<String>,
-    pub error: Option<RemoteError>,
-}
-
-/// Structured remote-side error envelope.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct RemoteError {
-    pub code: String,
-    pub message: String,
-    #[serde(default)]
-    pub details: HashMap<String, serde_json::Value>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-/// Deterministic security-envelope validation failures for remote execution.
-///
-/// Use [`SecurityEnvelopeError::code`] as the stable classifier in tests and
-/// machine-readable handling.
-pub enum SecurityEnvelopeError {
-    /// Request was unsigned while receiver policy requires signatures.
-    ///
-    /// Recovery: attach a valid request signature or relax policy intentionally.
-    UnsignedRequestRequired,
-    /// Request omitted key_id while trust policy requires it.
-    ///
-    /// Recovery: provide a non-empty key_id in signature/security metadata.
-    MissingKeyId,
-    /// Signature algorithm is unsupported for remote request verification.
-    ///
-    /// Recovery: use a receiver-allowed algorithm (currently ed25519).
-    UnsupportedRequestSignatureAlgorithm { alg: String },
-    /// Signing is required but no request signature payload was provided.
-    ///
-    /// Recovery: sign canonical request bytes and include the signature payload.
-    MissingRequestSignature,
-    /// Signature payload is malformed and cannot be parsed/validated.
-    ///
-    /// Recovery: regenerate signature metadata and ensure schema/encoding is
-    /// correct.
-    MalformedRequestSignature { reason: String },
-    /// Signature verification failed against canonical request bytes.
-    ///
-    /// Recovery: re-sign the exact request payload without post-sign mutation.
-    RequestSignatureMismatch,
-    /// Request used a signature algorithm not allowed by receiver trust policy.
-    ///
-    /// Recovery: select an allowed algorithm or update receiver policy.
-    DisallowedAlgorithm { alg: String },
-    /// Request used a key source not allowed by receiver trust policy.
-    ///
-    /// Recovery: use an allowed key source (embedded/explicit_key as configured).
-    DisallowedKeySource { key_source: String },
-    /// A signed request omitted key-source metadata when policy requires it.
-    ///
-    /// Recovery: provide deterministic key-source metadata with the request.
-    MissingKeySource,
-    /// Requested path contains traversal semantics rejected by the envelope gate.
-    ///
-    /// Recovery: use sandbox-relative normalized paths.
-    PathTraversal { path: String },
-    /// Requested path could not be resolved because the sandbox target/root was missing.
-    ///
-    /// Recovery: create the sandbox root or the relevant ancestor under it.
-    PathNotFound { path: String },
-    /// Requested path could not be canonicalized in a deterministic/safe way.
-    ///
-    /// Recovery: normalize the sandbox root/path and retry.
-    PathNotCanonical { path: String },
-    /// Requested path was rejected because the sandbox policy disallows symlink traversal.
-    ///
-    /// Recovery: remove the symlinked component or change policy intentionally.
-    SymlinkDisallowed { path: String },
-    /// Requested path resolves outside sandbox root via symlink traversal.
-    ///
-    /// Recovery: constrain symlink targets to sandbox root.
-    SymlinkEscape { path: String },
-    /// Requested path failed sandbox validation due to a filesystem I/O error.
-    ///
-    /// Recovery: inspect filesystem permissions/state and retry.
-    SandboxIoError {
-        path: String,
-        operation: &'static str,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-/// Stable remote execute client-side error kinds.
-///
-/// This enum normalizes transport/protocol/remote failures for deterministic
-/// handling and retry classification.
-pub enum RemoteExecuteClientErrorKind {
-    /// Request timed out before a valid response was received.
-    Timeout,
-    /// Remote endpoint was unreachable.
-    Unreachable,
-    /// Remote endpoint returned a non-success HTTP status.
-    BadStatus,
-    /// Response body could not be parsed as expected JSON schema.
-    InvalidJson,
-    /// Request payload violated remote execution schema constraints.
-    SchemaViolation,
-    /// Remote endpoint executed request but returned a structured execution error.
-    RemoteExecution,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct RemoteExecuteClientError {
-    pub kind: RemoteExecuteClientErrorKind,
-    pub code: String,
-    pub message: String,
-}
-
-impl std::fmt::Display for RemoteExecuteClientError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}: {}", self.code, self.message)
-    }
-}
-
-impl std::error::Error for RemoteExecuteClientError {}
-
-impl RemoteExecuteClientError {
-    pub(crate) fn new(
-        kind: RemoteExecuteClientErrorKind,
-        code: impl Into<String>,
-        message: impl Into<String>,
-    ) -> Self {
-        Self {
-            kind,
-            code: code.into(),
-            message: message.into(),
-        }
-    }
-}
-
-pub fn retryability(err: &anyhow::Error) -> Option<bool> {
-    for cause in err.chain() {
-        if cause.downcast_ref::<SecurityEnvelopeError>().is_some() {
-            return Some(false);
-        }
-        if let Some(remote) = cause.downcast_ref::<RemoteExecuteClientError>() {
-            return Some(match remote.kind {
-                RemoteExecuteClientErrorKind::Timeout
-                | RemoteExecuteClientErrorKind::Unreachable
-                | RemoteExecuteClientErrorKind::BadStatus
-                | RemoteExecuteClientErrorKind::InvalidJson => true,
-                RemoteExecuteClientErrorKind::SchemaViolation => false,
-                RemoteExecuteClientErrorKind::RemoteExecution => {
-                    !matches!(
-                        remote.code.as_str(),
-                        "REMOTE_SCHEMA_VIOLATION"
-                            | "SIGN_POLICY_UNSIGNED_REQUIRED"
-                            | "SIGN_POLICY_MISSING_KEY_ID"
-                            | "SIGN_POLICY_DISALLOWED_ALGORITHM"
-                            | "SIGN_POLICY_DISALLOWED_KEY_SOURCE"
-                            | "SIGN_POLICY_MISSING_KEY_SOURCE"
-                    ) && !remote.code.starts_with("REMOTE_ENVELOPE_")
-                }
-            });
-        }
-    }
-    None
-}
-
-pub fn stable_failure_kind(err: &anyhow::Error) -> Option<&'static str> {
-    for cause in err.chain() {
-        if cause.downcast_ref::<SecurityEnvelopeError>().is_some() {
-            return Some("policy_denied");
-        }
-        if let Some(remote) = cause.downcast_ref::<RemoteExecuteClientError>() {
-            return Some(match remote.kind {
-                RemoteExecuteClientErrorKind::Timeout => "timeout",
-                RemoteExecuteClientErrorKind::SchemaViolation => "schema_error",
-                RemoteExecuteClientErrorKind::Unreachable
-                | RemoteExecuteClientErrorKind::BadStatus
-                | RemoteExecuteClientErrorKind::InvalidJson => "io_error",
-                RemoteExecuteClientErrorKind::RemoteExecution => "provider_error",
-            });
-        }
-    }
-    None
-}
-
-impl SecurityEnvelopeError {
-    pub fn code(&self) -> &'static str {
-        match self {
-            Self::UnsignedRequestRequired => "REMOTE_ENVELOPE_UNSIGNED_REQUEST",
-            Self::MissingKeyId => "REMOTE_ENVELOPE_MISSING_KEY_ID",
-            Self::UnsupportedRequestSignatureAlgorithm { .. } => {
-                "REMOTE_REQUEST_SIGNATURE_UNSUPPORTED_ALGORITHM"
-            }
-            Self::MissingRequestSignature => "REMOTE_REQUEST_SIGNATURE_MISSING",
-            Self::MalformedRequestSignature { .. } => "REMOTE_REQUEST_SIGNATURE_MALFORMED",
-            Self::RequestSignatureMismatch => "REMOTE_REQUEST_SIGNATURE_MISMATCH",
-            Self::DisallowedAlgorithm { .. } => "REMOTE_ENVELOPE_DISALLOWED_ALGORITHM",
-            Self::DisallowedKeySource { .. } => "REMOTE_ENVELOPE_DISALLOWED_KEY_SOURCE",
-            Self::MissingKeySource => "REMOTE_ENVELOPE_MISSING_KEY_SOURCE",
-            Self::PathTraversal { .. } => "REMOTE_ENVELOPE_PATH_TRAVERSAL",
-            Self::PathNotFound { .. } => "REMOTE_ENVELOPE_PATH_NOT_FOUND",
-            Self::PathNotCanonical { .. } => "REMOTE_ENVELOPE_PATH_NOT_CANONICAL",
-            Self::SymlinkDisallowed { .. } => "REMOTE_ENVELOPE_SYMLINK_DISALLOWED",
-            Self::SymlinkEscape { .. } => "REMOTE_ENVELOPE_SYMLINK_ESCAPE",
-            Self::SandboxIoError { .. } => "REMOTE_ENVELOPE_SANDBOX_IO_ERROR",
-        }
-    }
-
-    pub fn message(&self) -> String {
-        match self {
-            Self::UnsignedRequestRequired => {
-                "remote security envelope rejected unsigned request while signing is required"
-                    .to_string()
-            }
-            Self::MissingKeyId => {
-                "remote security envelope requires non-empty key_id when trust policy is enabled"
-                    .to_string()
-            }
-            Self::UnsupportedRequestSignatureAlgorithm { alg } => format!(
-                "remote request signing rejected unsupported signature algorithm '{alg}' (expected '{REMOTE_REQUEST_SIGNATURE_ALG_ED25519}')"
-            ),
-            Self::MissingRequestSignature => {
-                "remote request signing required signature payload but none was provided"
-                    .to_string()
-            }
-            Self::MalformedRequestSignature { reason } => {
-                format!("remote request signature is malformed: {reason}")
-            }
-            Self::RequestSignatureMismatch => {
-                "remote request signature verification failed (canonical request mismatch)"
-                    .to_string()
-            }
-            Self::DisallowedAlgorithm { alg } => format!(
-                "remote security envelope rejected signature algorithm '{alg}' per verification profile"
-            ),
-            Self::DisallowedKeySource { key_source } => format!(
-                "remote security envelope rejected key source '{key_source}' per verification profile"
-            ),
-            Self::MissingKeySource => {
-                "remote security envelope rejected request: missing key source for signature verification policy".to_string()
-            }
-            Self::PathTraversal { path } => format!(
-                "remote security envelope rejected requested path with traversal/absolute components: '{path}'"
-            ),
-            Self::PathNotFound { path } => format!(
-                "remote security envelope rejected requested path because the sandbox target/root was not found: '{path}'"
-            ),
-            Self::PathNotCanonical { path } => format!(
-                "remote security envelope rejected requested path because sandbox canonicalization failed deterministically: '{path}'"
-            ),
-            Self::SymlinkDisallowed { path } => format!(
-                "remote security envelope rejected requested path because symlink traversal is disabled by sandbox policy: '{path}'"
-            ),
-            Self::SymlinkEscape { path } => format!(
-                "remote security envelope rejected requested path escaping sandbox root via symlink/canonicalization: '{path}'"
-            ),
-            Self::SandboxIoError { path, operation } => format!(
-                "remote security envelope rejected requested path because sandbox validation hit an IO error during {operation}: '{path}'"
-            ),
-        }
-    }
-}
-
-impl std::fmt::Display for SecurityEnvelopeError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}: {}", self.code(), self.message())
-    }
-}
-
-impl std::error::Error for SecurityEnvelopeError {}
-
-fn sort_value(value: &mut Value) {
-    match value {
-        Value::Object(map) => {
-            let mut sorted = std::collections::BTreeMap::new();
-            for (k, mut v) in std::mem::take(map) {
-                sort_value(&mut v);
-                sorted.insert(k, v);
-            }
-            let mut out = Map::new();
-            for (k, v) in sorted {
-                out.insert(k, v);
-            }
-            *map = out;
-        }
-        Value::Array(items) => {
-            for item in items {
-                sort_value(item);
-            }
-        }
-        _ => {}
-    }
-}
-
-pub fn canonical_request_bytes(req: &ExecuteRequest) -> Result<Vec<u8>> {
-    // Canonicalization contract (v1):
-    // 1) clone request
-    // 2) exclude only `security.request_signature`
-    // 3) recursively sort JSON object keys (including HashMap-backed fields)
-    // 4) serialize compact JSON bytes
-    //
-    // Client and server both use this routine for signature verification.
-    // Any change here is a wire-compatibility change for signed requests.
-    let mut canonical = req.clone();
-    if let Some(sec) = canonical.security.as_mut() {
-        sec.request_signature = None;
-    }
-    let mut value = serde_json::to_value(canonical)
-        .context("failed to convert execute request to canonical JSON")?;
-    sort_value(&mut value);
-    serde_json::to_vec(&value).context("failed to serialize canonical execute request")
-}
-
-pub fn sign_execute_request_v1(
-    req: &ExecuteRequest,
-    private_key_b64: &str,
-    key_id: Option<&str>,
-) -> Result<RemoteRequestSignatureV1> {
-    let key_bytes = B64
-        .decode(private_key_b64.trim().as_bytes())
-        .context("invalid base64 private key for remote request signing")?;
-    let key_arr: [u8; 32] = key_bytes
-        .try_into()
-        .map_err(|_| anyhow!("remote request private key must be exactly 32 bytes"))?;
-    let signing = SigningKey::from_bytes(&key_arr);
-    let canonical = canonical_request_bytes(req)?;
-    let sig = signing.sign(&canonical);
-    Ok(RemoteRequestSignatureV1 {
-        schema_version: REMOTE_REQUEST_SIGNATURE_SCHEMA_V1.to_string(),
-        alg: REMOTE_REQUEST_SIGNATURE_ALG_ED25519.to_string(),
-        key_id: key_id.map(|v| v.to_string()),
-        public_key_b64: Some(B64.encode(signing.verifying_key().to_bytes())),
-        sig_b64: B64.encode(sig.to_bytes()),
-    })
-}
-
-fn attach_request_signature(
-    req: &mut ExecuteRequest,
-    private_key_b64: &str,
-    key_id: Option<&str>,
-) -> Result<()> {
-    {
-        let env = req
-            .security
-            .get_or_insert_with(ExecuteSecurityEnvelope::default);
-        env.signed = true;
-        if let Some(id) = key_id {
-            env.key_id = Some(id.to_string());
-        }
-        env.signature_alg = Some(REMOTE_REQUEST_SIGNATURE_ALG_ED25519.to_string());
-        env.key_source = Some("embedded".to_string());
-    }
-
-    // Canonical request bytes include envelope metadata and exclude only the
-    // request_signature payload itself.
-    let signature = sign_execute_request_v1(req, private_key_b64, key_id)?;
-    let env = req
-        .security
-        .get_or_insert_with(ExecuteSecurityEnvelope::default);
-    env.request_signature = Some(signature);
-    Ok(())
-}
-
-pub fn maybe_attach_request_signature_from_env(req: &mut ExecuteRequest) -> Result<()> {
-    let require_signature = req
-        .security
-        .as_ref()
-        .map(|env| env.require_signature)
-        .unwrap_or(false);
-    let private_key = std::env::var("ADL_REMOTE_REQUEST_SIGNING_PRIVATE_KEY_B64")
-        .ok()
-        .map(|v| v.trim().to_string())
-        .filter(|v| !v.is_empty());
-    let key_id = std::env::var("ADL_REMOTE_REQUEST_SIGNING_KEY_ID")
-        .ok()
-        .map(|v| v.trim().to_string())
-        .filter(|v| !v.is_empty())
-        .or_else(|| req.security.as_ref().and_then(|env| env.key_id.clone()));
-
-    match (require_signature, private_key) {
-        (true, None) => {
-            return Err(anyhow!(
-                "REMOTE_REQUEST_SIGNATURE_MISSING: signing is required but ADL_REMOTE_REQUEST_SIGNING_PRIVATE_KEY_B64 is not set"
-            ));
-        }
-        (_, None) => return Ok(()),
-        (_, Some(private_key_b64)) => {
-            attach_request_signature(req, &private_key_b64, key_id.as_deref())?;
-        }
-    }
-    Ok(())
-}
-
-fn verify_execute_request_signature_v1(
-    req: &ExecuteRequest,
-    sig: &RemoteRequestSignatureV1,
-) -> std::result::Result<(), SecurityEnvelopeError> {
-    if sig.schema_version.trim() != REMOTE_REQUEST_SIGNATURE_SCHEMA_V1 {
-        return Err(SecurityEnvelopeError::MalformedRequestSignature {
-            reason: format!(
-                "unexpected schema_version '{}' (expected '{}')",
-                sig.schema_version, REMOTE_REQUEST_SIGNATURE_SCHEMA_V1
-            ),
-        });
-    }
-    if !sig
-        .alg
-        .trim()
-        .eq_ignore_ascii_case(REMOTE_REQUEST_SIGNATURE_ALG_ED25519)
-    {
-        return Err(
-            SecurityEnvelopeError::UnsupportedRequestSignatureAlgorithm {
-                alg: sig.alg.clone(),
-            },
-        );
-    }
-    let pub_b64 = sig.public_key_b64.as_deref().ok_or_else(|| {
-        SecurityEnvelopeError::MalformedRequestSignature {
-            reason: "missing public_key_b64".to_string(),
-        }
-    })?;
-    let pub_bytes = B64.decode(pub_b64.as_bytes()).map_err(|_| {
-        SecurityEnvelopeError::MalformedRequestSignature {
-            reason: "invalid base64 public_key_b64".to_string(),
-        }
-    })?;
-    let pub_arr: [u8; 32] =
-        pub_bytes
-            .try_into()
-            .map_err(|_| SecurityEnvelopeError::MalformedRequestSignature {
-                reason: "public key must be exactly 32 bytes".to_string(),
-            })?;
-    let public = VerifyingKey::from_bytes(&pub_arr).map_err(|_| {
-        SecurityEnvelopeError::MalformedRequestSignature {
-            reason: "invalid ed25519 public key".to_string(),
-        }
-    })?;
-    let sig_bytes = B64.decode(sig.sig_b64.as_bytes()).map_err(|_| {
-        SecurityEnvelopeError::MalformedRequestSignature {
-            reason: "invalid base64 sig_b64".to_string(),
-        }
-    })?;
-    let parsed_sig = Signature::from_slice(&sig_bytes).map_err(|_| {
-        SecurityEnvelopeError::MalformedRequestSignature {
-            reason: "invalid ed25519 signature bytes".to_string(),
-        }
-    })?;
-    let canonical = canonical_request_bytes(req).map_err(|err| {
-        SecurityEnvelopeError::MalformedRequestSignature {
-            reason: format!("failed to canonicalize request: {err:#}"),
-        }
-    })?;
-    public
-        .verify(&canonical, &parsed_sig)
-        .map_err(|_| SecurityEnvelopeError::RequestSignatureMismatch)?;
-    Ok(())
-}
-
-pub fn validate_security_envelope(
-    req: &ExecuteRequest,
-) -> std::result::Result<(), SecurityEnvelopeError> {
-    let env = req.security.as_ref().cloned().unwrap_or_default();
-    let signature_payload = env.request_signature.clone();
-    if env.require_signature && signature_payload.is_none() {
-        return Err(SecurityEnvelopeError::MissingRequestSignature);
-    }
-
-    let derived_alg = signature_payload
-        .as_ref()
-        .map(|sig| sig.alg.clone())
-        .or_else(|| env.signature_alg.clone());
-    let derived_key_id = signature_payload
-        .as_ref()
-        .and_then(|sig| sig.key_id.clone())
-        .or_else(|| env.key_id.clone());
-    let derived_key_source = if signature_payload
-        .as_ref()
-        .and_then(|sig| sig.public_key_b64.as_ref())
-        .is_some()
-    {
-        Some(signing::VerificationKeySource::Embedded)
-    } else {
-        match env.key_source.as_deref() {
-            Some(raw) => match signing::VerificationKeySource::parse(raw) {
-                Some(source) => Some(source),
-                None => {
-                    return Err(SecurityEnvelopeError::DisallowedKeySource {
-                        key_source: raw.to_string(),
-                    });
-                }
-            },
-            None => None,
-        }
-    };
-    let requested_algs = env
-        .allowed_algs
-        .iter()
-        .map(|raw| raw.trim().to_ascii_lowercase())
-        .filter(|value| !value.is_empty())
-        .collect::<Vec<_>>();
-    let requested_key_sources = env
-        .allowed_key_sources
-        .iter()
-        .map(|raw| {
-            signing::VerificationKeySource::parse(raw).ok_or_else(|| {
-                SecurityEnvelopeError::DisallowedKeySource {
-                    key_source: raw.clone(),
-                }
-            })
-        })
-        .collect::<std::result::Result<Vec<_>, _>>()?;
-    let mut receiver_profile = signing::VerificationProfile::default().canonicalized();
-    let allowed_algs = if requested_algs.is_empty() {
-        receiver_profile.allowed_algs.clone()
-    } else {
-        for alg in &requested_algs {
-            if !receiver_profile.allowed_algs.iter().any(|base| base == alg) {
-                return Err(SecurityEnvelopeError::DisallowedAlgorithm { alg: alg.clone() });
-            }
-        }
-        requested_algs
-    };
-    let allowed_key_sources = if requested_key_sources.is_empty() {
-        receiver_profile.allowed_key_sources.clone()
-    } else {
-        for source in &requested_key_sources {
-            if !receiver_profile.allowed_key_sources.contains(source) {
-                return Err(SecurityEnvelopeError::DisallowedKeySource {
-                    key_source: source.as_str().to_string(),
-                });
-            }
-        }
-        requested_key_sources
-    };
-    receiver_profile.require_signature = env.require_signature;
-    receiver_profile.require_key_id = env.require_key_id;
-    receiver_profile.allowed_algs = allowed_algs;
-    receiver_profile.allowed_key_sources = allowed_key_sources;
-    let profile = signing::VerificationProfile {
-        require_signature: receiver_profile.require_signature,
-        require_key_id: receiver_profile.require_key_id,
-        allowed_algs: receiver_profile.allowed_algs.clone(),
-        allowed_key_sources: receiver_profile.allowed_key_sources.clone(),
-    };
-    let metadata = signing::VerificationMetadata {
-        signed: signature_payload.is_some() || env.signed,
-        key_id: derived_key_id.as_deref(),
-        alg: derived_alg.as_deref(),
-        key_source: derived_key_source,
-    };
-    if let Err(err) = signing::enforce_verification_profile(&metadata, &profile) {
-        let mapped = match err.code {
-            "SIGN_POLICY_UNSIGNED_REQUIRED" => SecurityEnvelopeError::UnsignedRequestRequired,
-            "SIGN_POLICY_MISSING_KEY_ID" => SecurityEnvelopeError::MissingKeyId,
-            "SIGN_POLICY_DISALLOWED_ALGORITHM" => SecurityEnvelopeError::DisallowedAlgorithm {
-                alg: derived_alg.as_deref().unwrap_or("<unknown>").to_string(),
-            },
-            "SIGN_POLICY_DISALLOWED_KEY_SOURCE" => SecurityEnvelopeError::DisallowedKeySource {
-                key_source: env
-                    .key_source
-                    .as_deref()
-                    .or_else(|| derived_key_source.as_ref().map(|source| source.as_str()))
-                    .unwrap_or("<unknown>")
-                    .to_string(),
-            },
-            "SIGN_POLICY_MISSING_KEY_SOURCE" => SecurityEnvelopeError::MissingKeySource,
-            _ => SecurityEnvelopeError::MissingKeySource,
-        };
-        return Err(mapped);
-    }
-    // Enforce trust policy constraints first, then verify cryptographic integrity.
-    // This ensures algorithm/key_id/key_source gating fails deterministically
-    // before any signature-byte verification path is evaluated.
-    if let Some(sig) = signature_payload.as_ref() {
-        verify_execute_request_signature_v1(req, sig)?;
-    }
-
-    if env.requested_paths.is_empty() {
-        return Ok(());
-    }
-
-    let sandbox_root = env.sandbox_root.as_deref().unwrap_or(".");
-    for rel in &env.requested_paths {
-        let resolved = sandbox::resolve_relative_path_for_write_within_root(
-            std::path::Path::new(sandbox_root),
-            std::path::Path::new(rel),
-        );
-        if let Err(err) = resolved {
-            return Err(map_sandbox_path_error(rel, &err));
-        }
-    }
-    Ok(())
-}
-
-fn map_sandbox_path_error(
-    requested_path: &str,
-    err: &sandbox::SandboxPathError,
-) -> SecurityEnvelopeError {
-    match err {
-        sandbox::SandboxPathError::PathDenied { .. } => SecurityEnvelopeError::PathTraversal {
-            path: requested_path.to_string(),
-        },
-        sandbox::SandboxPathError::PathNotFound { .. } => SecurityEnvelopeError::PathNotFound {
-            path: requested_path.to_string(),
-        },
-        sandbox::SandboxPathError::PathNotCanonical { .. } => {
-            SecurityEnvelopeError::PathNotCanonical {
-                path: requested_path.to_string(),
-            }
-        }
-        sandbox::SandboxPathError::SymlinkDisallowed { .. } => {
-            SecurityEnvelopeError::SymlinkDisallowed {
-                path: requested_path.to_string(),
-            }
-        }
-        sandbox::SandboxPathError::EscapeAttempt { .. } => SecurityEnvelopeError::SymlinkEscape {
-            path: requested_path.to_string(),
-        },
-        sandbox::SandboxPathError::IoError { operation, .. } => {
-            SecurityEnvelopeError::SandboxIoError {
-                path: requested_path.to_string(),
-                operation,
-            }
-        }
-    }
-}
-
-impl ExecuteResponse {
-    pub fn ok(req: &ExecuteRequest, output: String) -> Self {
-        Self {
-            ok: true,
-            run_id: req.run_id.clone(),
-            workflow_id: req.workflow_id.clone(),
-            step_id: req.step_id.clone(),
-            result: Some(output),
-            artifacts: Vec::new(),
-            error: None,
-        }
-    }
-
-    pub fn err(req: &ExecuteRequest, code: &str, message: impl Into<String>) -> Self {
-        Self {
-            ok: false,
-            run_id: req.run_id.clone(),
-            workflow_id: req.workflow_id.clone(),
-            step_id: req.step_id.clone(),
-            result: None,
-            artifacts: Vec::new(),
-            error: Some(RemoteError {
-                code: code.to_string(),
-                message: message.into(),
-                details: HashMap::new(),
-            }),
-        }
-    }
-}
+pub use self::errors::{
+    retryability, stable_failure_kind, RemoteExecuteClientError, RemoteExecuteClientErrorKind,
+    SecurityEnvelopeError,
+};
+#[cfg(test)]
+use self::security::map_sandbox_path_error;
+pub use self::security::validate_security_envelope;
+#[cfg(test)]
+use self::signing_support::attach_request_signature;
+pub use self::signing_support::{
+    canonical_request_bytes, maybe_attach_request_signature_from_env, sign_execute_request_v1,
+};
+pub use self::types::{
+    ExecuteInputsPayload, ExecuteRequest, ExecuteResponse, ExecuteSecurityEnvelope,
+    ExecuteStepPayload, RemoteError, RemoteRequestSignatureV1,
+};
 
 /// Execute one step against a remote execution endpoint.
 ///
